+++ client/src/game_state.rs
@@ -1,11 +1,15 @@
+use std::collections::HashMap;
+use std::rc::Rc;
+
 use ragnarok_formats::act::ActFile;
 use ragnarok_formats::gat::GatFile;
 use ragnarok_game::accessory_table::AccessoryTable;
 use ragnarok_game::app_state::AppState;
 use ragnarok_game::cursor::CursorAnimationState;
-use ragnarok_game::entity::Entity;
-use ragnarok_game::event::CharacterInfo;
+use ragnarok_game::entity_collection::EntityCollection;
 use ragnarok_game::map_coordinates::MapCoordinates;
+use ragnarok_game::name_table::NameTable;
+use ragnarok_game::event::CharacterInfo;
 use ragnarok_network::session::Session;
 use ragnarok_renderer::{EntitySprite, SpriteTextures};
 
@@ -16,8 +20,10 @@ pub struct GameState {
     pub current_map: Option<String>,
     pub map_coords: Option<MapCoordinates>,
     pub gat: Option<GatFile>,
-    pub player_entity: Option<Entity>,
-    pub player_sprite: Option<EntitySprite>,
+    pub entities: EntityCollection,
+    pub sprites: HashMap<u32, Rc<EntitySprite>>,
+    pub sprite_cache: HashMap<String, Rc<EntitySprite>>,
+    pub name_table: Option<NameTable>,
     pub accessory_table: Option<AccessoryTable>,
     pub cursor_textures: Option<SpriteTextures>,
     pub cursor_act: Option<ActFile>,
@@ -33,8 +39,10 @@ impl GameState {
             current_map: None,
             map_coords: None,
             gat: None,
-            player_entity: None,
-            player_sprite: None,
+            entities: EntityCollection::new(),
+            sprites: HashMap::new(),
+            sprite_cache: HashMap::new(),
+            name_table: None,
             accessory_table: None,
             cursor_textures: None,
             cursor_act: None,