+++ lib/formats/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod act;
 pub mod builtin_accessory_table;
+pub mod builtin_name_table;
 pub mod lua_table;
 pub mod gat;
 pub mod gnd;